Moves some shared utils to server utils
