+++ server/database/queries/notification.js
@@ -58,7 +58,7 @@ export const toggleNotification = async (id, isEnabled = true) => {
 
 export const deleteNotification = async (id) => {
   await SQLite.client('notifications').where({ id }).del();
-  await SQLite.client('notifications')
+  await SQLite.client('monitor')
     .where({ id })
     .update({ notificationId: null });
   return;
