+++ server/middleware/auth/login.js
@@ -1,10 +1,8 @@
 // import local files
 import { signInUser } from '../../database/queries/user.js';
 import { setServerSideCookie } from '../../../shared/utils/cookies.js';
-import {
-  handleError,
-  UnprocessableError,
-} from '../../../shared/utils/errors.js';
+import { handleError } from '../../utils/errors.js';
+import { UnprocessableError } from '../../../shared/utils/errors.js';
 import validators from '../../../shared/validators/index.js';
 
 const login = async (request, response) => {
