+++ server/middleware/auth/logout.js
@@ -1,5 +1,5 @@
 import { deleteCookie } from '../../../shared/utils/cookies.js';
-import { handleError } from '../../../shared/utils/errors.js';
+import { handleError } from '../../utils/errors.js';
 import { createURL } from '../../../shared/utils/url.js';
 
 const logout = (request, response) => {
