+++ server/middleware/monitor/delete.js
@@ -1,9 +1,7 @@
 // import local files
 import cache from '../../cache/index.js';
-import {
-  UnprocessableError,
-  handleError,
-} from '../../../shared/utils/errors.js';
+import { handleError } from '../../utils/errors.js';
+import { UnprocessableError } from '../../../shared/utils/errors.js';
 
 const monitorDelete = async (request, response) => {
   try {
