+++ server/middleware/monitor/edit.js
@@ -1,8 +1,6 @@
 // import local files
-import {
-  handleError,
-  UnprocessableError,
-} from '../../../shared/utils/errors.js';
+import { handleError } from '../../utils/errors.js';
+import { UnprocessableError } from '../../../shared/utils/errors.js';
 import validators from '../../../shared/validators/monitor.js';
 import cache from '../../cache/index.js';
 import { userExists } from '../../database/queries/user.js';
