+++ server/middleware/monitor/id.js
@@ -1,9 +1,7 @@
 import cache from '../../cache/index.js';
 import { cleanMonitor } from '../../class/monitor.js';
-import {
-  UnprocessableError,
-  handleError,
-} from '../../../shared/utils/errors.js';
+import { handleError } from '../../utils/errors.js';
+import { UnprocessableError } from '../../../shared/utils/errors.js';
 
 const fetchMonitorUsingId = async (request, response) => {
   try {
