+++ server/middleware/user/hasEditor.js
@@ -1,5 +1,5 @@
 import { userExists } from '../../database/queries/user.js';
-import { handleError } from '../../../shared/utils/errors.js';
+import { handleError } from '../../utils/errors.js';
 
 const hasEditorPermissions = async (request, response, next) => {
   try {
