+++ server/middleware/user/team/members.js
@@ -1,5 +1,5 @@
 import { fetchMembers } from '../../../database/queries/user.js';
-import { handleError } from '../../../../shared/utils/errors.js';
+import { handleError } from '../../../utils/errors.js';
 
 const teamMembersListMiddleware = async (request, response) => {
   try {
