+++ server/middleware/user/transferOwnership.js
@@ -3,7 +3,7 @@ import {
   transferOwnership,
   userExists,
 } from '../../database/queries/user.js';
-import { handleError } from '../../../shared/utils/errors.js';
+import { handleError } from '../../utils/errors.js';
 
 const transferOwnershipMiddleware = async (request, response) => {
   try {
