+++ server/middleware/user/update/username.js
@@ -2,7 +2,7 @@ import {
   updateUserDisplayname,
   userExists,
 } from '../../../database/queries/user.js';
-import { handleError } from '../../../../shared/utils/errors.js';
+import { handleError } from '../../../utils/errors.js';
 import validators from '../../../../shared/validators/index.js';
 
 const userUpdateUsername = async (request, response) => {
