+++ server/tools/tcpPing.js
@@ -2,7 +2,7 @@
 import net from 'net';
 
 // import local files
-import logger from '../../shared/utils/logger.js';
+import logger from '../utils/logger.js';
 
 const tcpStatusCheck = async (monitor, callback) => {
   const socket = new net.Socket();
@@ -18,10 +18,9 @@ const tcpStatusCheck = async (monitor, callback) => {
     const latency = Date.now() - startTime;
     socket.destroy();
 
-    logger.error(
-      'TCP Status Check',
-      `Issue checking monitor ${monitor.monitorId}: TIMED OUT`
-    );
+    logger.error('TCP Status Check', {
+      message: `Issue checking monitor ${monitor.monitorId}: TIMED OUT`,
+    });
 
     callback(monitor, {
       monitorId: monitor.monitorId,
@@ -49,10 +48,9 @@ const tcpStatusCheck = async (monitor, callback) => {
     const latency = Date.now() - startTime;
     socket.destroy();
 
-    logger.error(
-      'TCP Status Check',
-      `Issue checking monitor ${monitor.monitorId}: ${error.message} `
-    );
+    logger.error('TCP Status Check', {
+      message: `Issue checking monitor ${monitor.monitorId}: ${error.message}`,
+    });
 
     callback(monitor, {
       monitorId: monitor.monitorId,
