+++ server/utils/cron.js
@@ -2,12 +2,12 @@
 import { CronJob } from 'cron';
 
 // import local files
-import cache from '../../server/cache/index.js';
-import logger from './logger.js';
+import cache from '../cache/index.js';
+import logger from '../utils/logger.js';
 import {
   fetchHeartbeatsByDate,
   fetchLastDailyHeartbeat,
-} from '../../server/database/queries/heartbeat.js';
+} from '../database/queries/heartbeat.js';
 
 // fetch all monitors
 // fetch only heartbeats that are up for each monitor
@@ -15,12 +15,15 @@ import {
 // create a new record in the database
 
 async function initialiseCronJobs() {
-  logger.info('Cron', 'Initialising cron jobs');
+  logger.info('Cron', { message: 'Initialising cron jobs' });
 
+  // Evert hour
   new CronJob(
     '0 * * * *',
     async function () {
-      logger.info('Cron', 'Running hourly cron job for creating heartbeat');
+      logger.info('Cron', {
+        message: 'Running hourly cron job for creating heartbeat',
+      });
 
       const monitors = await cache.monitors.getKeys();
 
@@ -71,7 +74,9 @@ async function initialiseCronJobs() {
   new CronJob(
     '*/5 * * * *',
     async function () {
-      logger.info('Cron', 'Running 5 minute cron job for fetching heartbeats');
+      logger.info('Cron', {
+        message: 'Running 5 minute cron job for fetching heartbeats',
+      });
 
       const monitors = await cache.monitors.getKeys();
