+++ server/utils/hashPassword.js
